chemotracker/requirements.txt
appdirs==1.4.3
certifi==2017.11.5
chardet==3.0.4
Django==1.11.1
django-allauth==0.34.0
django-rest-auth==0.9.2
django-watson==1.5.2
djangorestframework==3.6.3
idna==2.6
oauthlib==2.0.6
packaging==16.8
psycopg2==2.7.4
pyparsing==2.2.0
python-openid==2.2.5
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.10.0
urllib3==1.22
virtualenv==15.1.0
virtualenvwrapper-win==1.2.1
